List the contents of the Java distributions cache folder

List the contents of the Java distributions cache folder

Usage

java_list_distrib_cache(
  output = c("data.frame", "vector"),
  verbose = FALSE,
  cache_dir = tools::R_user_dir("rJavaEnv", which = "cache")
)

Arguments

output

The format of the output: "data.frame" or "vector". Defaults to "data.frame".

verbose

Whether to print detailed messages. Defaults to FALSE.

cache_dir

The cache directory to list. Defaults to the user-specific data directory.

Value

A character vector with the contents of the cache directory.

Examples

java_list_distrib_cache()
#> ✖ Path does not exist
#> character(0)